backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LangChain & LangGraph
langchain>=0.1.0
langgraph>=0.0.26
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0

# Document Processing
pymupdf>=1.23.0
pdf2image>=1.16.0
Pillow>=10.0.0

# Vector Database
qdrant-client>=1.7.0

# ColPali for visual document embeddings
colpali-engine>=0.2.0
torch>=2.0.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
